refactor(*) change the way services are split
