+++ src/views/Entities/ExternalServices.vue
@@ -1,106 +1,13 @@
 <template>
-  <div class="external-services">
-    <FrameSkeleton>
-      <DataOverview
-        :page-size="pageSize"
-        :has-error="hasError"
-        :is-loading="isLoading"
-        :empty-state="empty_state"
-        :display-data-table="true"
-        :table-data="tableData"
-        :table-data-is-empty="tableDataIsEmpty"
-        table-data-function-text="View"
-        table-data-row="name"
-        @tableAction="tableAction"
-        @reloadData="loadData"
-      >
-        <template slot="additionalControls">
-          <KButton
-            v-if="this.$route.query.ns"
-            class="back-button"
-            appearance="primary"
-            size="small"
-            :to="{
-              name: 'external-services'
-            }"
-          >
-            <span class="custom-control-icon">
-              &larr;
-            </span>
-            View All
-          </KButton>
-        </template>
-        <template slot="pagination">
-          <Pagination
-            :has-previous="previous.length > 0"
-            :has-next="hasNext"
-            @next="goToNextPage"
-            @previous="goToPreviousPage"
-          />
-        </template>
-      </DataOverview>
-      <Tabs
-        v-if="isEmpty === false"
-        :has-error="hasError"
-        :is-loading="isLoading"
-        :tabs="tabs"
-        initial-tab-override="overview"
-      >
-        <template slot="tabHeader">
-          <div>
-            <h3>{{ tabGroupTitle }}</h3>
-          </div>
-          <div>
-            <EntityURLControl :url="shareUrl" />
-          </div>
-        </template>
-        <template slot="overview">
-          <LabelList
-            :has-error="entityHasError"
-            :is-loading="entityIsLoading"
-            :is-empty="entityIsEmpty"
-          >
-            <div>
-              <ul>
-                <li
-                  v-for="(val, key) in entity"
-                  :key="key"
-                >
-                  <h4>{{ key }}</h4>
-                  <p>
-                    {{ val }}
-                  </p>
-                </li>
-              </ul>
-            </div>
-          </LabelList>
-        </template>
-        <template slot="yaml">
-          <YamlView
-            lang="yaml"
-            :title="entityOverviewTitle"
-            :has-error="entityHasError"
-            :is-loading="entityIsLoading"
-            :is-empty="entityIsEmpty"
-            :content="rawEntity"
-          />
-        </template>
-      </Tabs>
-    </FrameSkeleton>
-  </div>
+  <Services
+    route-name="external-services"
+    name="External Services"
+    :tab-headers="tabHeaders"
+  />
 </template>
-
 <script>
-import { getOffset, getSome, stripTimes } from '@/helpers'
-import EntityURLControl from '@/components/Utils/EntityURLControl'
-import sortEntities from '@/mixins/EntitySorter'
-import FormatForCLI from '@/mixins/FormatForCLI'
-import FrameSkeleton from '@/components/Skeletons/FrameSkeleton'
-import Pagination from '@/components/Pagination'
-import DataOverview from '@/components/Skeletons/DataOverview'
-import Tabs from '@/components/Utils/Tabs'
-import YamlView from '@/components/Skeletons/YamlView'
-import LabelList from '@/components/Utils/LabelList'
+
+import Services from '@/views/Entities/partial/Services.vue'
 
 export default {
   name: 'ExternalServices',
@@ -108,266 +15,17 @@ export default {
     title: 'External Services'
   },
   components: {
-    EntityURLControl,
-    FrameSkeleton,
-    Pagination,
-    DataOverview,
-    Tabs,
-    YamlView,
-    LabelList
+    Services,
   },
-  mixins: [
-    FormatForCLI,
-    sortEntities
-  ],
   data () {
     return {
-      isLoading: true,
-      isEmpty: false,
-      hasError: false,
-      entityIsLoading: true,
-      entityIsEmpty: false,
-      entityHasError: false,
-      tableDataIsEmpty: false,
-      empty_state: {
-        title: 'No Data',
-        message: 'There are no External Services present.'
-      },
-      tableData: {
-        headers: [
-          { key: 'actions', hideLabel: true },
-          { label: 'Name', key: 'name' },
-          { label: 'Mesh', key: 'mesh' },
-          { label: 'Address', key: 'address' },
-          { label: 'TLS', key: 'tlsEnabled' }
-        ],
-        data: []
-      },
-      tabs: [
-        {
-          hash: '#overview',
-          title: 'Overview'
-        },
-        {
-          hash: '#yaml',
-          title: 'YAML'
-        }
-      ],
-      entity: [],
-      rawEntity: null,
-      firstEntity: null,
-      pageSize: this.$pageSize,
-      pageOffset: null,
-      next: null,
-      hasNext: false,
-      previous: []
-    }
-  },
-  computed: {
-    tabGroupTitle () {
-      const entity = this.entity
-
-      if (entity) {
-        return `External Service: ${entity.name}`
-      } else {
-        return null
-      }
-    },
-    entityOverviewTitle () {
-      const entity = this.entity
-
-      if (entity) {
-        return `Entity Overview for ${entity.name}`
-      } else {
-        return null
-      }
-    },
-    formattedRawEntity () {
-      const entity = this.formatForCLI(this.rawEntity)
-
-      return entity
-    },
-    shareUrl () {
-      const urlRoot = `${window.location.origin}#`
-      const entity = this.entity
-
-      const shareUrl = () => {
-        if (this.$route.query.ns) {
-          return this.$route.fullPath
-        }
-
-        return `${urlRoot}${this.$route.fullPath}?ns=${entity.name}`
-      }
-
-      return shareUrl()
-    }
-  },
-  watch: {
-    '$route' (to, from) {
-      this.init()
-    }
-  },
-  beforeMount () {
-    this.init()
-  },
-  methods: {
-    init () {
-      this.loadData()
-    },
-    goToPreviousPage () {
-      this.pageOffset = this.previous.pop()
-      this.next = null
-
-      this.loadData()
-    },
-    goToNextPage () {
-      this.previous.push(this.pageOffset)
-      this.pageOffset = this.next
-      this.next = null
-
-      this.loadData()
-    },
-    tableAction (ev) {
-      const data = ev
-
-      // load the data into the tabs
-      this.getEntity(data)
-    },
-    loadData () {
-      this.isLoading = true
-
-      const mesh = this.$route.params.mesh || null
-      const query = this.$route.query.ns || null
-
-      const params = {
-        size: this.pageSize,
-        offset: this.pageOffset
-      }
-
-      const endpoint = () => {
-        if (mesh === 'all') {
-          return this.$api.getAllExternalServices(params)
-        } else if ((query && query.length) && mesh !== 'all') {
-          return this.$api.getExternalService(mesh, query, params)
-        }
-
-        return this.$api.getAllExternalServicesFromMesh(mesh)
-      }
-
-      const getExternalServices = () => endpoint()
-        .then(response => {
-          const items = () => {
-            const r = response
-
-            if ('total' in r) {
-              if (r.total !== 0 && r.items && r.items.length > 0) {
-                return this.sortEntities(r.items)
-              }
-
-              return null
-            }
-
-            return r
-          }
-
-          const entityList = items()
-
-          if (items()) {
-            const firstItem = query
-              ? entityList
-              : entityList[0]
-
-            // set the first item as the default for initial load
-            this.firstEntity = firstItem.name
-
-            // load the YAML entity for the first item on page load
-            this.getEntity(stripTimes(firstItem))
-
-            this.tableData.data = query
-              ? [entityList]
-              : entityList
-
-            if (response.next) {
-              this.next = getOffset(response.next)
-              this.hasNext = true
-            } else {
-              this.hasNext = false
-            }
-
-            this.tableData.data = this.tableData.data.map(entity => {
-              const { networking = {} } = entity
-              const { tls = {} } = networking
-
-              entity.address = networking.address
-              entity.tlsEnabled = tls.enabled ? 'Enabled' : 'Disabled'
-
-              return entity
-            })
-
-            this.tableDataIsEmpty = false
-            this.isEmpty = false
-          } else {
-            this.tableData.data = []
-            this.tableDataIsEmpty = true
-            this.isEmpty = true
-
-            this.getEntity(null)
-          }
-        })
-        .catch(error => {
-          this.hasError = true
-          this.isEmpty = true
-
-          console.error(error)
-        })
-        .finally(() => {
-          setTimeout(() => {
-            this.isLoading = false
-          }, process.env.VUE_APP_DATA_TIMEOUT)
-        })
-
-      getExternalServices()
-    },
-    getEntity (entity) {
-      this.entityIsLoading = true
-      this.entityIsEmpty = false
-      this.entityHasError = false
-
-      const mesh = this.$route.params.mesh
-
-      if (entity && entity !== null) {
-        const entityMesh = (mesh === 'all')
-          ? entity.mesh
-          : mesh
-
-        return this.$api.getExternalService(entityMesh, entity.name)
-          .then(response => {
-            if (response) {
-              const selected = ['type', 'name', 'mesh']
-
-              this.entity = getSome(response, selected)
-              // this.rawEntity = response
-              this.rawEntity = stripTimes(response)
-            } else {
-              this.entity = null
-              this.entityIsEmpty = true
-            }
-          })
-          .catch(error => {
-            this.entityHasError = true
-            console.error(error)
-          })
-          .finally(() => {
-            setTimeout(() => {
-              this.entityIsLoading = false
-            }, process.env.VUE_APP_DATA_TIMEOUT)
-          })
-      } else {
-        setTimeout(() => {
-          this.entityIsEmpty = true
-          this.entityIsLoading = false
-        }, process.env.VUE_APP_DATA_TIMEOUT)
-      }
+      tabHeaders: [
+        { key: 'actions', hideLabel: true },
+        { label: 'Name', key: 'name' },
+        { label: 'Mesh', key: 'mesh' },
+        { label: 'Address', key: 'address' },
+        { label: 'TLS', key: 'tlsEnabled' }
+      ]
     }
   }
 }
